[feat/#14] remote Service 적용 및 변경
