+++ app/src/main/java/com/example/miniyoutube/data/repository/YoutubeRepository.kt
@@ -1,0 +1,9 @@
+package com.example.miniyoutube.data.repository
+
+import com.example.miniyoutube.data.model.remote.YoutubeVideo
+
+interface YoutubeRepository {
+    suspend fun requestSearch(q: String, videoCategoryId : String) : YoutubeVideo
+
+    suspend fun requestVideo(q: String) : YoutubeVideo
+}
